Broke out some functions for clarity.
